+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add Cat Editor seed input [#545](https://github.com/hushenghao/AndroidEasterEggs/issues/545)
 - Add to favorites in Cat Editor
 - Add Bengali translation, Thanks for [@ChildishRifty7](https://github.com/childishrifty7) 's contribution
+- Rocket Launcher add icon collection selection [#554](https://github.com/hushenghao/AndroidEasterEggs/issues/554)
 - Cat Editor bottom menu bar optimization
 - Fix Cat Editor draw problem on Android N [#37138664](https://issuetracker.google.com/issues/37138664)
 - Fix Cat Editor color hvs palette maxWidth [#522](https://github.com/hushenghao/AndroidEasterEggs/issues/522)